refactor: extract admin sqlite persistence lane
