+++ crates/dcc-mcp-gateway-admin/Cargo.toml
@@ -15,6 +15,9 @@ workspace = true
 [dependencies]
 axum.workspace = true
 chrono = { version = "0.4", features = ["serde"] }
+dcc-mcp-db = { path = "../dcc-mcp-db", default-features = false, optional = true }
+dcc-mcp-models.workspace = true
+dcc-mcp-transport.workspace = true
 parking_lot.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -24,6 +27,7 @@ workspace-hack = { version = "0.1", path = "../workspace-hack" }
 [features]
 default = []
 embed = []
+persist-sqlite = ["dep:dcc-mcp-db", "dcc-mcp-db/gateway-admin-sqlite"]
 
 [build-dependencies]
 dunce = "1.0"
